requirements.txt
pytest
fastapi
uvicorn
psycopg2-binary